fix(pg): decode SELECT 1 scope guard as i32, not i64
The owner-scope guard in replace_rule_targets decoded `SELECT 1`
as `(i64,)`. In PostgreSQL an integer literal is INT4, so sqlx
rejected it with "Rust type i64 (INT8) is not compatible with INT4",
surfacing as a generic "database error" when creating a forward rule
under Owner scope. SQLite's dynamic typing masked the bug, so it only
affected PostgreSQL deployments.

Decode as `(i32,)` to match the INT4 literal, consistent with the
identical guards in pg_repo/users.rs.

+++ crates/panel/src/db/pg_repo/rules.rs
@@ -96,7 +96,7 @@ impl RuleRepository for PgRepository {
     ) -> Result<(), DbError> {
         // Scope guard: under Owner scope, no-op unless the rule is owned by uid.
         if let Some(uid) = scope.owner_id() {
-            let owned: Option<(i64,)> =
+            let owned: Option<(i32,)> =
                 sqlx::query_as("SELECT 1 FROM forward_rules WHERE id = $1 AND uid = $2")
                     .bind(rule_id)
                     .bind(uid)
